Re-structured the code
Has to be tested
